git-remote-ipfs: Concurrent fetch/push
This is rather naive for now, as we're essentially running every IPFS API call
in its own Haskell thread. Observed speedups are significant nevertheless.

Upper bounds are placed on the number of concurrent connections to the API, as
well as the number of OS threads (via the `-maxN` RTS option).