git-remote-ipfs: Remove unnecessary dagPut after fetching LOB
The purpose of this seemed to be to validate that the IPFS object comprising a
LOB hashes to the expected value. We can achive the same without incurring
another API call by just writing the object to the git store and checking the
resulting hash against the requested one. Note that it does no harm to write a
corrupt object to the git store: it will be considered dangling and removed
when the next GC occurs.